- Rename packages - Update clock to have a clock face color - Update clock track to cast a shadow - Update dial to have a colored stroke base on track color - Remove unused tests - Commit idea config files to maintain code style etc - Move Time Change Listener to own interface file, like in v1.0.0 of fork branch that was in Java
