+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="picker.ugurtekbas.com.DialTimePicker" >
+    package="io.flyingmongoose.EtzioDemo" >
 
     <application
         android:allowBackup="false"
@@ -8,7 +8,7 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme" >
         <activity
-            android:name="picker.ugurtekbas.com.DialTimePicker.Main"
+            android:name="io.flyingmongoose.EtzioDemo.Main"
             android:label="@string/app_name" >
             <intent-filter>
                <action android:name="android.intent.action.MAIN" />
