+++ app/src/main/java/io/flyingmongoose/EtzioDemo/Main.kt
@@ -1,4 +1,4 @@
-package picker.ugurtekbas.com.DialTimePicker
+package io.flyingmongoose.EtzioDemo
 
 import android.os.Bundle
 import android.view.Menu
