+++ app/src/main/res/layout/ampm_picker.xml
@@ -3,6 +3,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:background="#C85757"
     android:orientation="vertical">
 
     <CheckBox
@@ -10,18 +11,14 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_gravity="center_horizontal"
-        android:checked="true"
-        android:text="@string/pickerEnable"
         android:layout_margin="5dp"
-        />
+        android:checked="true"
+        android:text="@string/pickerEnable" />
 
-    <picker.ugurtekbas.com.Picker.Picker
+    <io.flyingmongoose.EtzioTimePicker.Picker
         android:id="@+id/amPicker"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:layout_centerInParent="true"
-        android:layout_gravity="center_horizontal"
-        app:hourFormat="false"
-        />
+        android:layout_gravity="center_horizontal"/>
 
 </LinearLayout>
