+++ app/src/main/res/layout/normal_picker.xml
@@ -6,11 +6,10 @@
     android:orientation="vertical"
     >
 
-    <picker.ugurtekbas.com.Picker.Picker
+    <io.flyingmongoose.EtzioTimePicker.Picker
         android:id="@+id/picker"
         android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_centerInParent="true"
+        android:layout_height="wrap_content"
         android:layout_weight="0.8"
         android:layout_gravity="center_horizontal"
         />
