+++ app/src/main/res/menu/menu_main.xml
@@ -1,5 +1,5 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools" tools:context=".Main">
+    xmlns:tools="http://schemas.android.com/tools" tools:context="io.flyingmongoose.EtzioDemo.Main">
 
 </menu>
