+++ library/src/main/AndroidManifest.xml
@@ -1,2 +1,2 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="picker.ugurtekbas.com.library" />
+<manifest package="io.flyingmongoose.EtzioTimePicker.library" />
