+++ library/src/main/java/io/flyingmongoose/EtzioTimePicker/TimeChangedListener.kt
@@ -1,0 +1,11 @@
+package io.flyingmongoose.EtzioTimePicker
+
+import java.util.*
+
+/**
+ * @author Ugur Tekbas
+ */
+interface TimeChangedListener
+{
+    fun timeChanged(date: Date?)
+}
