+++ library/src/main/res/values/attrs.xml
@@ -2,40 +2,21 @@
 <resources>
 
     <declare-styleable name="Picker">
-        <attr
-            name="canvasColor"
-            format="color"
-            />
+        <attr name="canvasColor" format="color" />
 
-        <attr
-            name="clockColor"
-            format="color"
-            />
+        <attr name="clockColor" format="color" />
 
-        <attr
-            name="textColor"
-            format="color"
-            />
+        <attr name="clockFaceColor" format="color" />
 
-        <attr
-            name="dialColor"
-            format="color"
-            />
+        <attr name="textColor" format="color" />
 
-        <attr
-            name="hourFormat"
-            format="boolean"
-            />
+        <attr name="dialColor" format="color" />
 
-        <attr
-            name="trackSize"
-            format="dimension"
-            />
+        <attr name="hourFormat" format="boolean" />
 
-        <attr
-            name="dialRadius"
-            format="dimension"
-            />
+        <attr name="trackSize" format="dimension" />
+
+        <attr name="dialRadius" format="dimension" />
 
     </declare-styleable>
